Очищает точку входа и вносит правки после консультации
